finished OAuth & updated database for userId
