+++ client/src/script.js
@@ -1,31 +1,51 @@
 //ClientID: '42c01af939954a35a024a9d4aee4b125'
 //Redirect_URI: 'http://localhost:8080/api/callback'
 //Scope: 'user-read-private user-read-email playlist-modify-public playlist-modify-private'
-// const path = require('path');
-// const express = require('express');
-// const cors = require('cors') 
-// app.use(cors());
-
 
 const clientId = "42c01af939954a35a024a9d4aee4b125"; // Replace with your client ID
 const params = new URLSearchParams(window.location.search);
 const code = params.get("code");
 
+function getCookie(name) {
+  // document.cookie.match returns an arr of regex, name, and boolean
+  const cookieValue = document.cookie.match(
+    // this regex does the following
+    // ^ matches the beginning of the string
+    // |[^;]+ matches any character that is not a semicolon
+    // \s* matches any whitespace
+    // + matches the previous token 1 or more times
+    // \s* matches any whitespace
+    // = matches the = character
+    // \s* matches any whitespace
+    // ([^;]+) matches any character that is not a semicolon 1 or more times
+    // the whole thing is wrapped in parentheses to create a capture group
+      '(^|[^;]+)\\s*' + name + '\\s*=\\s*([^;]+)'
+  );
+  return cookieValue ? cookieValue.pop() : '';
+}
+
 if (!code) {
     redirectToAuthCodeFlow(clientId);
     console.log('we are in the first if statement')
 } else {
-  console.log('we are in the else statement')
+    console.log('we are in the else statement')
+
     const accessToken = await getAccessToken(clientId, code);
-    document.cookie = `Token=${accessToken}`
-    // // this is from Spotify's example
-    const profile = await fetchProfile(accessToken);
-    // console.log('profile data', profile);
-    // console.log('this is the accesstoken', accessToken)
+    const getCookieToken = getCookie('Token')
+
+    if (accessToken !== undefined) {
+      document.cookie = `Token=${accessToken}`
+    }
+    
     // GRAB USER ID FROM THE PROFILE RETURN OBJECT
-    // const user = profile.........
-    document.cookie = `User=${'1240934213'}`
-  }
+    const profile = await fetchProfile(accessToken);
+    const getCookieUser = getCookie('User')
+    if (profile.id !== undefined || getCookieUser === '') {
+      document.cookie = `User=${profile.id}`
+      // add a location reload 
+      location.reload();
+    }
+}
 
 async function redirectToAuthCodeFlow(clientId) {
   const verifier = generateCodeVerifier(128);
@@ -40,9 +60,6 @@ async function redirectToAuthCodeFlow(clientId) {
   params.append("scope", "user-read-private user-read-email playlist-modify-public playlist-modify-private");
   params.append("code_challenge_method", "S256");
   params.append("code_challenge", challenge);
-
-  // console.log('params below:')
-  // console.log(params.toString())
 
   document.location = `https://accounts.spotify.com/authorize?${params.toString()}`;
   console.log('after doc.location')
@@ -84,7 +101,6 @@ async function getAccessToken(clientId, code) {
   });
 
   const { access_token } = await result.json();
-  // console.log(access_token)
   return access_token;
   
 }
@@ -96,6 +112,7 @@ async function fetchProfile(token) {
   });
   // data received should be an object with an ID attribute
   await console.log('this should be an object with a lot of info on the user: ', result)
+  // save User
   const data = await result.json();
   console.log ('data', data)
   return data;
